Move project panel tests into their own file
